Move parts of GameService to PlayerService,  SquadService, & TileService
