+++ TomorrowDiesToday.Services/Game/ISquadService.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using TomorrowDiesToday.Models;
+
+namespace TomorrowDiesToday.Services.Game
+{
+    public interface ISquadService
+    {
+        IObservable<string> ErrorMessage { get; }
+
+        IObservable<Dictionary<string, SquadModel>> SelectedSquadsUpdate { get; }
+
+        IObservable<SquadModel> SquadUpdate { get; }
+
+        void CalculateSquadStats(SquadModel squadModel);
+
+        void ToggleSquad(SquadModel squadModel);
+    }
+}
